Move child components inside main Table component
