+++ src/components/Table/presenter.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { PropTypes } from 'prop-types';
-import Header from '../Header';
-import Item from '../Item';
+import Header from './Header';
+import Item from './Item';
 import conf from '../../config';
 
 const Table = ({ result }) => (
